+++ pom.xml
@@ -44,12 +44,13 @@
     <checkstyle.suppressions.location>src/conf/checkstyle/suppressions.xml</checkstyle.suppressions.location>
 
     <junit.version>4.12</junit.version>
-    <k3po.version>3.0.0-alpha-54</k3po.version>
+    <k3po.version>3.0.0-alpha-73</k3po.version>
+    <k3po.nukleus.ext.version>0.4</k3po.nukleus.ext.version>
     <jmh.version>1.12</jmh.version>
-    <nukleus.plugin.version>0.7.3</nukleus.plugin.version>
+    <nukleus.plugin.version>0.7.8</nukleus.plugin.version>
     <nukleus.version>0.2</nukleus.version>
     <nukleus.http.spec.version>0.6</nukleus.http.spec.version>
-    <nukleus.ws.spec.version>0.4</nukleus.ws.spec.version>
+    <nukleus.ws.spec.version>develop-SNAPSHOT</nukleus.ws.spec.version>
     <reaktor.test.version>0.1</reaktor.test.version>
     <reaktor.version>0.1</reaktor.version>
   </properties>
@@ -193,6 +194,13 @@
             </goals>
           </execution>
         </executions>
+        <dependencies>
+          <dependency>
+            <groupId>org.reaktivity</groupId>
+            <artifactId>k3po-nukleus-ext</artifactId>
+            <version>${k3po.nukleus.ext.version}</version>
+          </dependency>
+        </dependencies>
       </plugin>
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
